Move package ch.srg.pillarbox.core to ch.srgssr.pillarbox.core
