--- pillarbox-core-business/src/main/java/ch/srg/pillarbox/core/business/integrationlayer/data/Drm.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2022. SRG SSR. All rights reserved.
- * License information is available from the LICENSE file.
- */
-package ch.srg.pillarbox.core.business.integrationlayer.data
-
-import ch.srg.pillarbox.core.business.integrationlayer.data.Drm.Type.FAIRPLAY
-import com.squareup.moshi.JsonClass
-
-/**
- * Drm information
- *
- * @property type
- * @property licenseUrl
- * @property certificateUrl
- */
-@JsonClass(generateAdapter = true)
-data class Drm(val type: Type, val licenseUrl: String, val certificateUrl: String? = null) {
-
-    /**
-     * Drm Type, on Android [FAIRPLAY] is not playable!
-     */
-    enum class Type {
-        FAIRPLAY, WIDEVINE, PLAYREADY
-    }
-}
