--- pillarbox-core-business/src/main/java/ch/srg/pillarbox/core/business/tracker/SRGEventLoggerTracker.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2023. SRG SSR. All rights reserved.
- * License information is available from the LICENSE file.
- */
-package ch.srg.pillarbox.core.business.tracker
-
-import android.util.Log
-import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.exoplayer.util.EventLogger
-import ch.srgssr.pillarbox.player.tracker.MediaItemTracker
-
-/**
- * Enable/Disable EventLogger when item is currently active.
- */
-class SRGEventLoggerTracker : MediaItemTracker {
-    private val eventLogger = EventLogger(TAG)
-
-    override fun start(player: ExoPlayer) {
-        Log.w(TAG, "---- Start")
-        player.addAnalyticsListener(eventLogger)
-    }
-
-    override fun stop(player: ExoPlayer) {
-        Log.w(TAG, "---- Stop")
-        player.removeAnalyticsListener(eventLogger)
-    }
-
-    override fun update(data: Any) {
-        Log.w(TAG, "---- Update data = $data")
-    }
-
-    /**
-     * Factory for a [SRGEventLoggerTracker]
-     */
-    class Factory : MediaItemTracker.Factory {
-
-        override fun create(): MediaItemTracker {
-            return SRGEventLoggerTracker()
-        }
-    }
-
-    companion object {
-        private const val TAG = "SRGLogger"
-    }
-}
