+++ pillarbox-core-business/src/main/java/ch/srgssr/pillarbox/core/business/integrationlayer/service/RemoteResult.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2022. SRG SSR. All rights reserved.
+ * License information is available from the LICENSE file.
+ */
+package ch.srgssr.pillarbox.core.business.integrationlayer.service
+
+/**
+ * Remote result
+ *
+ * @param T type of the data
+ */
+sealed interface RemoteResult<T> {
+
+    /**
+     * Success
+     *
+     * @param T type of the data
+     * @property data
+     */
+    data class Success<T>(val data: T) : RemoteResult<T>
+
+    /**
+     * Error
+     *
+     * @param T type of the data, not used here
+     * @property throwable
+     * @property code
+     */
+    data class Error<T>(val throwable: Throwable, val code: Int = 404) : RemoteResult<T>
+}
